pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "sublist3r-vh"
version = "0.0.1"
description = "Subdomains enumeration tool for penetration testers and red teamings"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "dnspython",
    "requests",
    "rich",
    "pyyaml",
]

[project.scripts]
sublist3r = "sublist3r:interactive"

